requirements.txt
mysql-connector-python==8.0.23
numpy==1.19.5
pandas==1.2.1
Pillow==8.1.0
protobuf==3.14.0
pycodestyle==2.6.0
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
toml==0.10.2